Update game design, add new logo, buttons, checkboxes and connect World and Battle scenes to the template
