+++ src/Config/config.js
@@ -7,4 +7,12 @@ export default {
   parent: 'phaser-example',
   width: 800,
   height: 600,
+  pixelArt: true,
+  physics: {
+    default: 'arcade',
+    arcade: {
+      gravity: { y: 0 },
+      debug: true,
+    },
+  },
 };
