+++ src/Scenes/Battle.js
@@ -14,6 +14,7 @@ const BattleScene = new Phaser.Class({
   create() {
     // change the background to green
     this.cameras.main.setBackgroundColor('rgba(0, 200, 0, 0.5)');
+
     this.startBattle();
     // on wake event we call startBattle too
     this.sys.events.on('wake', this.startBattle, this);
