+++ src/Scenes/BootScene.js
@@ -8,7 +8,7 @@ export default class BootScene extends Phaser.Scene {
   }
 
   preload() {
-    this.load.image('logo', 'assets/zenva_logo.png');
+    this.load.image('rpg-logo', 'assets/rpg-logo.png');
   }
 
   create() {
