+++ src/Scenes/GameScene.js
@@ -14,5 +14,6 @@ export default class GameScene extends Phaser.Scene {
 
   create() {
     this.add.image(400, 300, 'logo');
+    this.scene.start('WorldScene');
   }
 }
