+++ src/Scenes/PreloaderScene.js
@@ -13,7 +13,7 @@ export default class PreloaderScene extends Phaser.Scene {
 
   preload() {
     // add logo image
-    this.add.image(400, 200, 'logo');
+    this.add.image(400, 150, 'rpg-logo').scale = 1.4;
 
     // display progress bar
     const progressBar = this.add.graphics();
@@ -82,12 +82,22 @@ export default class PreloaderScene extends Phaser.Scene {
     this.timedEvent = this.time.delayedCall(3000, this.ready, [], this);
 
     // load assets needed in our game
-    this.load.image('blueButton1', 'assets/ui/blue_button02.png');
-    this.load.image('blueButton2', 'assets/ui/blue_button03.png');
+    this.load.image('blueButton1', 'assets/ui/btn-brown2.png');
+    this.load.image('blueButton2', 'assets/ui/btn-brown1.png');
     this.load.image('phaserLogo', 'assets/logo.png');
-    this.load.image('box', 'assets/ui/grey_box.png');
-    this.load.image('checkedBox', 'assets/ui/blue_boxCheckmark.png');
+    this.load.image('box', 'assets/ui/unchecked.png');
+    this.load.image('checkedBox', 'assets/ui/checked.png');
     this.load.audio('bgMusic', ['assets/TownTheme.mp3']);
+
+    this.load.image('rpg-logo', 'assets/rpg-logo.png');
+
+    this.load.image('map', 'assets/map/rpg-map1.png');
+    this.load.image('tree', 'assets/map/tree.png');
+
+    this.load.spritesheet('player', 'assets/RPG_assets.png', { frameWidth: 16, frameHeight: 16 });
+
+    this.load.image('dragonblue', 'assets/monster-black.png');
+    this.load.image('dragonorrange', 'assets/monster-orange.png');
   }
 
   ready() {
