+++ src/index.js
@@ -9,6 +9,8 @@ import TitleScene from './Scenes/TitleScene';
 import OptionsScene from './Scenes/OptionsScene';
 import CreditsScene from './Scenes/CreditsScene';
 import Model from './Model';
+import WorldScene from './Scenes/WorldScene';
+import { BattleScene, UIScene } from './Scenes/Battle';
 
 class Game extends Phaser.Game {
   constructor() {
@@ -21,6 +23,9 @@ class Game extends Phaser.Game {
     this.scene.add('Options', OptionsScene);
     this.scene.add('Credits', CreditsScene);
     this.scene.add('Game', GameScene);
+    this.scene.add('WorldScene', WorldScene);
+    this.scene.add('BattleScene', BattleScene);
+    this.scene.add('UIScene', UIScene);
     this.scene.start('Boot');
   }
 }
